Revert "System protection added"
