+++ common/common.h
@@ -7,4 +7,3 @@
 #include "showPath/showPath.h"
 #include "copyFile/copyfile.h"
 #include "errorHandeling/errorHandeling.h"
-#include "searchInFile/searchInFile.h"
