--- common/searchInFile/searchInFile.h
@@ -1,3 +1,0 @@
-#include "../../mainGlobal.h"
-
-int searchInFile(char *filePath, char *str);
