+++ judge/judge.h
@@ -1,6 +1,10 @@
 #pragma once
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <dirent.h>
 #include <errno.h>
-#include "../mainGlobal.h"
+#include "../common/common.h"
 
 int compareFiles(const char path1[],const char path2[]);
 int checkFormatJudge (const char filePath[]);
